adding ewfinfo collection; fixing some formatting; changing MCT collection to keep the tmp file local in case of failures

+++ tsk-triage.sh
@@ -6,24 +6,35 @@ DISK="$1"
 NAME="$2"
 OUTPUTDIR="$3"
 if [ $# -eq 0 ]; then
-	USAGE "DISK" "OUTPUTDIR" && exit 1
+	USAGE "DISK" "NAME" "OUTPUTDIR" && exit 1
+	USAGE_DESCRIPTION "This script is meant to collect information about the filesystem(s) on a device and/or forensic image such that the analyst can get an idea of the volume/type/temporal data they're reviewing. Useful in making decisions on where to start a forensic analysis."
+	USAGE_EXAMPLE "$(basename "$0") /dev/rdisk8 \"EVIDENCE_ID_2\" ~/<case data>/triage/"
 fi
 
 LOG="$OUTPUTDIR/$NAME.log"
 START "$0" "$LOG" "$*"
 
-INFO "-----------------------------------------------------------------------" "$LOG"
-INFO "Display forensic operating system disk information (system_profiler)..." "$LOG"
-INFO "-----------------------------------------------------------------------" "$LOG"
-INFO "$(system_profiler SPUSBDataType SPParallelATADataType SPCardReaderDataType SPFireWireDataType SPHardwareRAIDDataType SPNVMeDataType SPParallelSCSIDataType SPSASDataType SPSerialATADataType SPThunderboltDataType | grep -B 14 -A 4 "BSD Name: $(basename "$DISK" | $SEDCMD -r 's/^r//')$" | $SEDCMD -r 's/^[[:space:]]*//; /^$/d')" "$LOG" | tee "$OUTPUTDIR/$NAME-system_profiler.txt"
-INFO "" "$LOG"
+if [ ! -f "$DISK" ]; then
+	INFO "-----------------------------------------------------------------------" "$LOG"
+	INFO "Display forensic operating system disk information (system_profiler)..." "$LOG"
+	INFO "-----------------------------------------------------------------------" "$LOG"
+	INFO "$(system_profiler SPUSBDataType SPParallelATADataType SPCardReaderDataType SPFireWireDataType SPHardwareRAIDDataType SPNVMeDataType SPParallelSCSIDataType SPSASDataType SPSerialATADataType SPThunderboltDataType | grep -B 14 -A 4 "BSD Name: $(basename "$DISK" | $SEDCMD -r 's/^r//')$" | $SEDCMD -r 's/^[[:space:]]*//; /^$/d')" "$LOG" | tee "$OUTPUTDIR/$NAME-system_profiler.txt"
+	INFO "" "$LOG"
 
-INFO "-----------------------------------------------------------------------" "$LOG"
-INFO "Display forensic operating system disk information (diskutil)..." "$LOG"
-INFO "-----------------------------------------------------------------------" "$LOG"
-INFO "$(diskutil info "$DISK")" "$LOG" | tee "$OUTPUTDIR/$NAME-diskutil.txt"
-INFO "$(diskutil list "$DISK")" "$LOG" | tee -a "$OUTPUTDIR/$NAME-diskutil.txt"
-INFO "" "$LOG"
+	INFO "-----------------------------------------------------------------------" "$LOG"
+	INFO "Display forensic operating system disk information (diskutil)..." "$LOG"
+	INFO "-----------------------------------------------------------------------" "$LOG"
+	INFO "$(diskutil info "$DISK")" "$LOG" | tee "$OUTPUTDIR/$NAME-diskutil.txt"
+	INFO "$(diskutil list "$DISK")" "$LOG" | tee -a "$OUTPUTDIR/$NAME-diskutil.txt"
+	INFO "" "$LOG"
+else
+	INFO "-----------------------------------------------------------------------" "$LOG"
+	INFO "Show meta data stored in EWF files (ewfinfo)..." "$LOG"
+	INFO "-----------------------------------------------------------------------" "$LOG"
+	LOG_VERSION "ewfinfo" "$(ewfinfo -V)" "$LOG"
+	INFO "$(ewfinfo "$DISK")" "$LOG" | tee "$OUTPUTDIR/$NAME-ewfinfo.txt"
+	INFO "" "$LOG"
+fi
 
 INFO "-----------------------------------------------------------------------" "$LOG"
 INFO "Display the partition layout of a volume system (mmls)..." "$LOG"
@@ -50,18 +61,19 @@ for PARTITION in $(_tsk_mmls_partitions "$DISK"); do
 done
 INFO "" "$LOG"
 
-MCT_TMP=$(MKTEMP "$0" || exit 1)
 INFO "-----------------------------------------------------------------------" "$LOG"
 INFO "Collect MAC times from a disk image into a body file (tsk_gettimes)..." "$LOG"
-INFO "$MCT_TMP" "$LOG"
 INFO "-----------------------------------------------------------------------" "$LOG"
 LOG_VERSION "tsk_gettimes" "$(tsk_gettimes -V)" "$LOG"
 LOG_VERSION "datatime" "$(datatime --version)" "$LOG"
-tsk_gettimes "$DISK" > "$MCT_TMP"
-"${BASH_SOURCE%/*}/tsk-fix-mct.sh" "$MCT_TMP" "$NAME" | tee "$OUTPUTDIR/$NAME.mct" | datatime > "$OUTPUTDIR/$NAME-mct.txt"
+tsk_gettimes "$DISK" > "$OUTPUTDIR/$NAME.mct.tmp"
+"${BASH_SOURCE%/*}/tsk-fix-mct.sh" "$OUTPUTDIR/$NAME.mct.tmp" "$NAME" | tee "$OUTPUTDIR/$NAME.mct" | datatime > "$OUTPUTDIR/$NAME-mct.txt"
+if [ -e "$OUTPUTDIR/$NAME.mct" ]; then
+	# Only delete the .mct.tmp file if the reprocessing was successful.
+	rm "$OUTPUTDIR/$NAME.mct.tmp"
+fi
 head-tail.sh "$OUTPUTDIR/$NAME-mct.txt" 50
 INFO "" "$LOG"
-rm "$MCT_TMP"
 
 NOTIFY "Finished triage for $NAME ($DISK)!" "$0"
 END "$0" "$LOG"
